+++ src/components/Button/index.js
@@ -1,2 +1,3 @@
 export default from './Button';
 export ButtonSkeleton from 'carbon-components-react/lib/components/Button/Button.Skeleton';
+export { Button as CarbonButton } from 'carbon-components-react';
