+++ src/components/Toolbar/index.js
@@ -1,0 +1,7 @@
+export {
+  Toolbar,
+  ToolbarItem,
+  ToolbarTitle,
+  ToolbarOption,
+  ToolbarDivider,
+} from 'carbon-components-react';
